requirements.txt
requests
flask
flask-debugtoolbar
waitress
pywebview